+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes will be documented in this file.
 
 ### Changed
 
+* Move `RateLimiter` (from roq-server) (#259)
 * API changes needed to support drop-copy (#254)
 * Add notional fields to `ReferenceData` (#252)
 * Add new fields to `RateLimitTrigger` (#251)